Split off runtime-tests to separate folder
This was needed since we only want the JIT compiler to run once before
the entire set of tests that check the BTree implementation. Otherwise
the JIT would run once for each (property) test!

Maybe in the future the two directories can get merged again, but it
would probably involve something hacky like the following (which I chose
not to do at the moment):

```haskell
spec :: IO ()
spec = runIO $ jitCompile $ \ffi -> hspec $ do
  -- ...
```

(This will run the tests while the test tree is being built up, which is
unexpected.)