Don't generate redundant ifs in hot-code path.
Previously in the SearchF case of lowering the RA IR, the query was
based on all clauses. However this was wasteful since it did unneeded,
expensive loads + comparisons in the hot code path.

Now the lowering algorithm takes into account the original form of the
RA IR, and disregards all "Constrain" nodes nested directly in the
search, yielding more optimized code.