Fix NoElem edge case during index selection
This commit fixes an issue I encountered when trying out the wildcards
feature. It generated an incorrect index for a relation since before it
used all `ColumnIndex` variants to see which rows were being accessed to
compute an index with.

Now the algorithm only takes into account the actual equality
constraints.