Support creating new nodes in simplify step.
This commit upgrades the `Transform` type to generate fresh `NodeId`
values. This is useful for when you have transformations that need to
create new nodes on the fly.

On top of that, the `RewriteRule` and `RewriteRuleT` type synonyms are
provided to help with typing transformations. This helps with better
errors during development.

An example:

```haskell
transform :: Transform AST AST
transform =
  Transform $ flip evalStateT 0 . cata rewrite
  where
    -- This allows for a pass that keeps track of some additional state
    rewrite :: RewriteRuleT (StateT Int) AST
    rewrite = _  -- TODO implementation here
```

Without the type synonym on rewrite, the Haskell typechecker would give
much more vague type errors (since the type of `cata` is so general).