Improve performance due to templated btree member function
Previously, when doing a merge on the RA level, both the source and
destination btree had to be *exactly* the same type. This also implied
having the same index to sort elements by. Because of this, extra btrees
had to be kept in memory and updated during the runtime of a program.

Now that we have the possibility of "templated member functions" thanks
to the `TemplateT` monad transformer, we can instantiate a specialized
`insert_range` call right before we use it.