Rewrite semantic analysis in Eclair itself
This is an important step towards bootstrapping! We can now generate
Souffle code from Eclair, and create a "stage0" compiler that way. Then,
we can use the stage0 compiler to compile the Eclair code directly
(resulting in "stage1", or the final compiler). The stage1 compiler does
not depend on Souffle!

Some constructs were rewritten in an equivalent form for now (because
Eclair doesn't support some syntactic sugar yet), but they can be
reintroduced later. It shouldn't be any difference to the compiler
though, because the desugared form is what Souffle compiles it to
anyway.